tmux configuration (#30)
* tmux configuration

* update installation script

+++ install.sh
@@ -2,14 +2,37 @@
 
 SCRIPT_DIR=$(dirname "$(realpath "$0")")
 
-rm ~/.config/nvim
+
+if [[ -d ~/.config/nvim ]]; then
+  rm ~/.config/nvim
+fi
 ln -s "$SCRIPT_DIR/nvim" ~/.config/nvim
 
-rm ~/.config/bat
+if [[ -d ~/.config/bat ]]; then
+  rm ~/.config/bat
+fi
 ln -s "$SCRIPT_DIR/bat" ~/.config/bat
 
-rm ~/.local/bin/tmux-sessionizer
+if [[ ! -d ~/.config/tmux ]]; then
+  mkdir -p ~/.config/tmux
+fi
+if [[ -f ~/.config/tmux/tmux.conf ]]; then
+  rm ~/.config/tmux/tmux.conf
+fi
+ln -s "$SCRIPT_DIR/tmux/tmux.conf" ~/.config/tmux/tmux.conf
+
+if [[ ! -d ~/.local/bin ]]; then
+  mkdir -p ~/.local/bin
+fi
+if [[ -f ~/.local/bin/tmux-sessionizer ]]; then
+  rm ~/.local/bin/tmux-sessionizer
+fi
 ln -s "$SCRIPT_DIR/scripts/tmux-sessionizer" ~/.local/bin/tmux-sessionizer
 
-rm ~/.config/lazygit/config.yml
+if [[ ! -d ~/.config/lazygit ]]; then
+  mkdir -p ~/.config/lazygit
+fi
+if [[ -f ~/.config/lazygit/config.yml ]]; then
+  rm ~/.config/lazygit/config.yml
+fi
 ln -s "$SCRIPT_DIR/lazygit/config.yml" ~/.config/lazygit/config.yml
